introduced family.pomdp submodule; split the policy tree into two files one holding the tree structure and one holding the synthesizer
