+++ paynt/family/__init__.py
@@ -1,4 +1,5 @@
 from .colored_mdp import FamilyColoredMdp
 from .factory import FamilyColoredMdpFactory
-from .pomdp_family import PomdpFamilyColoredMdp, PomdpFamilyColoredMdpFactory
-from .policy_tree import PolicyTree, PolicyTreeNode, MdpFamilyResult, PolicyTreeSynthesizer
+from .pomdp import PomdpFamilyColoredMdp, PomdpFamilyColoredMdpFactory
+from .policy_tree import PolicyTree, PolicyTreeNode
+from .policy_tree_synthesizer import MdpFamilyResult, PolicyTreeSynthesizer
