+++ paynt/family/policy_tree_synthesizer.py
@@ -1,0 +1,300 @@
+'''
+AR-based synthesis of a policy tree (paynt.family.policy_tree.PolicyTree): splits a family's parameter space
+until every subspace either has a policy that satisfies all its members, or is proven unsatisfiable. Unlike
+the generic AR/CEGIS/Hybrid synthesizers, this never goes through SynthesizerAR -- policy-tree splitting is a
+genuinely different algorithm (game abstraction + policy compatibility merging), not a scoring variant.
+'''
+
+import paynt.synthesizer.synthesizer
+import paynt.underlying_model.underlying_model
+import paynt.utils.scoring
+from paynt.family.policy_tree import PolicyTree
+
+import logging
+logger = logging.getLogger(__name__)
+
+
+class MdpFamilyResult:
+    def __init__(self):
+        # if None, then family is undediced
+        # if False, then all family members are UNSAT
+        # otherwise, contains a satisfying policy for all MDPs in the family
+        self.policy = None
+
+        self.game_policy = None
+        self.parameter_selection = None
+        self.splitter = None
+
+class PolicyTreeSynthesizer(paynt.synthesizer.synthesizer.Synthesizer):
+
+    # if True, tree leaves will be double-checked after synthesis
+    double_check_policy_tree_leaves = False
+
+    @property
+    def method_name(self):
+        return "AR (policy tree)"
+
+    def verify_policy(self, selected_choices, prop, policy):
+        '''
+        :param selected_choices the compatible-choices bitmask to verify against -- callers outside this
+            class have no PolicyTreeNode to read it from, so it must be supplied explicitly (see
+            ParameterSpaceEvaluation.selected_choices, captured at decision time)
+        '''
+        _,mdp = self.colored_mdp.fix_and_apply_policy_to_parameter_space(selected_choices, policy)
+        policy_result = mdp.model_check_property(prop, alt=True)
+        self.stat.iteration(mdp)
+        return policy_result.sat
+
+
+    def solve_singleton(self, node, prop):
+        result = node.mdp.model_check_property(prop)
+        self.stat.iteration(node.mdp)
+        if not result.sat:
+            return False
+        policy = self.colored_mdp.scheduler_to_policy(result.result.scheduler, node.mdp)
+
+        # uncomment below to preemptively double-check the policy
+        # paynt.family.policy_tree.double_check_policy(self.colored_mdp, node, prop, policy)
+        return policy
+
+
+    def solve_game_abstraction(self, node, prop, game_solver):
+        # construct and solve the game abstraction
+        # logger.debug("solving game abstraction...")
+
+        game_solver.solve_sg(node.selected_choices)
+        # game_solver.solve_smg(node.selected_choices)
+
+        game_value = game_solver.solution_value
+        self.stat.iteration_game(node.mdp.states)
+        game_sat = prop.satisfies_threshold_within_precision(game_value)
+        # logger.debug("game solved, value is {}".format(game_value))
+        game_policy = game_solver.solution_state_to_player1_action
+        # fix irrelevant choices
+        game_policy_fixed = self.colored_mdp.empty_policy()
+        for state,action in enumerate(game_policy):
+            if action < self.colored_mdp.num_actions:
+                game_policy_fixed[state] = action
+        game_policy = game_policy_fixed
+        return game_policy,game_sat
+
+    def state_to_choice_to_parameter_selection(self, state_to_choice):
+        if self.task.discard_unreachable_choices:
+            state_to_choice = paynt.underlying_model.underlying_model.ModelIndex.discard_unreachable_choices(
+                self.colored_mdp.underlying_mdp, self.colored_mdp.choice_destinations, state_to_choice)
+        scheduler_choices = paynt.underlying_model.underlying_model.ModelIndex.state_to_choice_to_choices(
+            self.colored_mdp.underlying_mdp, state_to_choice)
+        parameter_selection = self.colored_mdp.coloring.collectHoleOptions(scheduler_choices)
+        return scheduler_choices,parameter_selection
+
+    def parse_game_scheduler(self, game_solver):
+        state_to_choice = game_solver.solution_state_to_quotient_choice.copy()
+        scheduler_choices,parameter_selection = self.state_to_choice_to_parameter_selection(state_to_choice)
+        state_values = game_solver.solution_state_values
+        return scheduler_choices,parameter_selection,state_values
+
+    def verify_parameter_space(self, node, game_solver, prop):
+        # logger.info("investigating parameter space of size {}".format(node.parameter_space.size))
+        node.mdp, node.selected_choices = self.colored_mdp.build(node.parameter_space)
+        mdp_family_result = MdpFamilyResult()
+
+        if node.parameter_space.size == 1:
+            mdp_family_result.policy = self.solve_singleton(node,prop)
+            return mdp_family_result
+
+        if node.candidate_policy is None:
+            game_policy,game_sat = self.solve_game_abstraction(node,prop,game_solver)
+        else:
+            game_policy = node.candidate_policy
+            game_sat = False
+
+        mdp_family_result.game_policy = game_policy
+        if game_sat:
+            mdp_family_result.policy = game_policy
+            return mdp_family_result
+
+        # solve primary direction for the MDP abstraction
+        mdp_result = node.mdp.model_check_property(prop)
+        mdp_value = mdp_result.value
+        self.stat.iteration(node.mdp)
+        # logger.debug("primary-primary direction solved, value is {}".format(mdp_value))
+        if not mdp_result.sat:
+            mdp_family_result.policy = False
+            return mdp_family_result
+
+        # undecided: choose scheduler choices to be used for splitting
+        scheduler_choices,parameter_selection,state_values = self.parse_game_scheduler(game_solver)
+
+        splitter = self.choose_splitter(node.parameter_space,prop,scheduler_choices,state_values,parameter_selection)
+        mdp_family_result.splitter = splitter
+        mdp_family_result.parameter_selection = parameter_selection
+        return mdp_family_result
+
+    def choose_splitter(self, parameter_space, prop, scheduler_choices, state_values, parameter_selection):
+        inconsistent_assignments = {parameter:options for parameter,options in enumerate(parameter_selection) if len(options) > 1}
+        if len(inconsistent_assignments)==0:
+            # pick any parameter with multiple options involved in the parameter selection
+            for parameter,options in enumerate(parameter_selection):
+                if parameter_space.parameter_num_options(parameter) > 1 and len(options) > 0:
+                    return parameter
+            # pick any parameter with multiple options
+            # logger.debug("picking an arbitrary parameter...")
+            for parameter in range(parameter_space.num_parameters):
+                if parameter_space.parameter_num_options(parameter) > 1:
+                    return parameter
+        if len(inconsistent_assignments)==1:
+            for parameter in inconsistent_assignments.keys():
+                return parameter
+
+        # compute scores for inconsistent parameters
+        scores = self.compute_scores(prop, scheduler_choices, state_values, inconsistent_assignments)
+        splitters = paynt.utils.scoring.parameters_with_max_score(scores)
+        splitter = splitters[0]
+        return splitter
+
+    def compute_scores(self, prop, scheduler_choices, state_values, inconsistent_assignments):
+        mdp = self.colored_mdp.underlying_mdp
+        choice_values = paynt.underlying_model.underlying_model.ModelIndex.choice_values(mdp, prop, state_values)
+        expected_visits = paynt.underlying_model.underlying_model.ModelIndex.compute_expected_visits(
+            mdp, prop, scheduler_choices, disable_expected_visits=self.task.disable_expected_visits)
+        underlying_mdp_choice_map = [choice for choice in range(self.colored_mdp.underlying_mdp.nr_choices)]
+        scores = paynt.utils.scoring.estimate_scheduler_difference(
+            self.colored_mdp, self.colored_mdp.underlying_mdp, underlying_mdp_choice_map, inconsistent_assignments, choice_values, expected_visits)
+        return scores
+
+    def assign_candidate_policy(self, parameter_subspaces, candidate_policies, parameter_selection, splitter, policy):
+        ''' Fill in candidate_policies[i] = policy for whichever parameter_subspaces[i] contains the parameter
+        selection, in place -- parameter_subspaces are plain ParameterSpace values here (not yet wrapped as
+        PolicyTreeNode children), so the candidate policy can't be attached directly until attach_children
+        wraps them (see evaluate_all). '''
+        policy_consistent = all([len(options) <= 1 for options in parameter_selection])
+        if not policy_consistent:
+            return
+        # associate the branch of the split that contains the parameter selection with the policy
+        used_options = parameter_selection[splitter]
+        if len(used_options) != 1:
+            # not sure what to do in this case
+            return
+        option = used_options[0]
+        for index,parameter_subspace in enumerate(parameter_subspaces):
+            if option in parameter_subspace.parameter_options(splitter):
+                candidate_policies[index] = policy
+                return
+
+    def split(self, parameter_space, prop, parameter_selection, splitter, policy):
+        # split the parameter
+        used_options = parameter_selection[splitter]
+        if len(used_options) > 1:
+            # used_options = used_options[0:1]
+            core_suboptions = [[option] for option in used_options]
+            other_suboptions = [option for option in parameter_space.parameter_options(splitter) if option not in used_options]
+            if other_suboptions:
+                other_suboptions = [other_suboptions]
+            else:
+                other_suboptions = []
+            suboptions = other_suboptions + core_suboptions # DFS solves core first
+        else:
+            options = parameter_space.parameter_options(splitter)
+            assert len(options) > 1
+            half = len(options) // 2
+            suboptions = [options[:half], options[half:]]
+
+        parameter_subspaces = parameter_space.split(splitter,suboptions)
+        candidate_policies = [None for _ in parameter_subspaces]
+
+        if not self.task.discard_unreachable_choices:
+            self.assign_candidate_policy(parameter_subspaces, candidate_policies, parameter_selection, splitter, policy)
+
+        return suboptions,parameter_subspaces,candidate_policies
+
+
+    def evaluate_all(self, parameter_space, prop, keep_value_only=False):
+        assert not prop.reward, "expecting reachability probability propery"
+        game_solver = self.colored_mdp.build_game_abstraction_solver(prop)
+        policy_tree = PolicyTree(parameter_space)
+
+        undecided_leaves = [policy_tree.root]
+        while undecided_leaves:
+
+            # gi = self.stat.iterations_game
+            # if gi is not None and gi > 1000:
+            #     return None
+
+            node = undecided_leaves.pop(-1)
+            result = self.verify_parameter_space(node,game_solver,prop)
+            node.candidate_policy = None
+
+            if result.policy is not None:
+                self.explore(node.parameter_space)
+                if node != policy_tree.root:
+                    node.mdp = None
+                if result.policy is False:
+                    node.sat = False
+                else:
+                    node.sat = True
+                    node.policy_index = policy_tree.new_policy(result.policy)
+                continue
+
+            # refine
+            suboptions,parameter_subspaces,candidate_policies = self.split(node.parameter_space, prop, result.parameter_selection, result.splitter, result.game_policy)
+            if node != policy_tree.root:
+                node.mdp = None
+            node.attach_children(result.splitter,suboptions,parameter_subspaces,candidate_policies)
+            undecided_leaves += node.child_nodes
+
+        if PolicyTreeSynthesizer.double_check_policy_tree_leaves:
+            policy_tree.double_check(self.colored_mdp, prop)
+        policy_tree.print_stats()
+
+        self.stat.num_mdps_total = self.colored_mdp.parameter_space.size
+        self.stat.num_mdps_sat = sum([n.parameter_space.size for n in policy_tree.collect_sat()])
+        self.stat.num_nodes = len(policy_tree.collect_all())
+        self.stat.num_leaves = len(policy_tree.collect_leaves())
+        self.stat.num_policies = len(policy_tree.policies)
+        postprocessing_time = policy_tree.postprocess(self.colored_mdp, prop)
+        policy_tree.print_stats()
+        self.stat.postprocessing_time = postprocessing_time
+        self.stat.num_nodes_merged = len(policy_tree.collect_all())
+        self.stat.num_leaves_merged = len(policy_tree.collect_leaves())
+        self.stat.num_policies_merged = len(policy_tree.policies)
+        self.policy_tree = policy_tree
+
+        # convert policy tree to parameter space evaluations
+        evaluations = []
+        for node in policy_tree.collect_leaves():
+            policy = policy_tree.policies[node.policy_index] if node.sat else None
+            evaluation = paynt.synthesizer.synthesizer.ParameterSpaceEvaluation(
+                node.parameter_space,None,node.sat,policy=policy,selected_choices=node.selected_choices)
+            evaluations.append(evaluation)
+        return evaluations
+
+
+    def run(self, optimum_threshold=None):
+        return self.evaluate()
+
+
+    def export_evaluation_result(self, evaluations, export_filename_base):
+        import json
+        policies = self.policy_tree.extract_policies(self.colored_mdp)
+        policies_json = {}
+        for index,key_value in enumerate(policies.items()):
+            policy_id,policy = key_value
+            policy_json = self.colored_mdp.policy_to_json(policy)
+            policies_json[policy_id] = policy_json
+        policies_string = json.dumps(policies_json, indent=4)
+
+        policies_filename = export_filename_base + ".json"
+        with open(policies_filename, 'w') as file:
+            file.write(policies_string)
+
+        logger.info(f"exported policies to {policies_filename}")
+
+        tree = self.policy_tree.extract_policy_tree(self.colored_mdp)
+        tree_filename = export_filename_base + ".dot"
+        with open(tree_filename, 'w') as file:
+            file.write(tree.source)
+        logger.info(f"exported policy tree to {tree_filename}")
+
+        tree_visualization_filename = export_filename_base + ".png"
+        tree.render(export_filename_base, format="png", cleanup=True) # using export_filename_base since graphviz appends .png by default
+        logger.info(f"exported policy tree visualization to {tree_visualization_filename}")
