+++ paynt/family/pomdp/__init__.py
@@ -1,0 +1,2 @@
+from .colored_mdp import PomdpFamilyColoredMdp, SubPomdp
+from .factory import PomdpFamilyColoredMdpFactory
